Added notepad creator
This tool will scan a repository of PDF files and create a txt file for each containing the cover page.

The intended purpose is the creation of places to take notes on research papers for things such as a lit review. It saves the effort of copying over information such as title, authors, date, abstract, etc.